feat: Create dummy components for views and add navigation template
This change set adds five template components. They are
1) DeckList: for displaying the list of decks
2) Deck: for displaying the details of an individual deck where
the user can start a quiz or add questions to the deck
3) NewCard: The user can add a new card (i.e. a question) to the
deck using this view
4) NewDeck: The user can create a new deck using this view
5) Quiz: The user can take the quiz using this view

The main view has two tabs, one for the deck list and the other one
for creating a new deck. The user will be able to navigate from deck
list to an individual deck and from deck to the quiz view or the view
for adding new cards. The template for this navigation is also set in
this change set